Develop a SKOS concept scheme with collections
Added information to make this a SKOS Concept scheme and associate individual
concepts with the scheme.

Add a set of "concept collections" <http://www.w3.org/TR/skos-primer/#seccollections>
to organize the individual rights statements into the three groups mentioned in
<https://docs.google.com/document/d/1x10JsIfi8Y74pgJJEAqMtyO5iYp0p6DO5DrOZK-5umY/preview?sle=true>.
Concept collections seems to be a better and looser structure than more
formal semantic relations (broader/narrower).  If these kinds of semantic
relations should be added, we need to discuss how to create the nodes that
will organize them.  For example, while we have a In-Copyright concept that
could be the top topic for more specific in-copyright statements, there are
no general "public domain" or "undetermined" concepts here.